ShareIt add-bookings plz

+++ src/main/java/ru/practicum/shareit/booking/model/BookingDate.java
@@ -3,9 +3,10 @@ package ru.practicum.shareit.booking.model;
 import java.time.LocalDateTime;
 
 public interface BookingDate {
-    Integer getId();
+    Long getId();
 
     LocalDateTime getBookingDate();
 
-    Integer getBookerId();
+    Long getBookerId();
+
 }
