+++ src/main/java/ru/practicum/shareit/item/dto/CommentDto.java
@@ -1,13 +1,17 @@
 package ru.practicum.shareit.item.dto;
 
+
 import lombok.Data;
 
 import java.time.LocalDateTime;
 
 @Data
 public class CommentDto {
-    private int id;
+    private Long id;
+
     private String text;
+
     private String authorName;
+
     private LocalDateTime created;
 }
